Develop

+++ server/src/seeds/cleanDb.ts
@@ -5,10 +5,12 @@ export default async (modelName: "Question", collectionName: string) => {
   try {
     let modelExists = await models[modelName]?.db?.db?.listCollections({
       name: collectionName
-    }).toArray()
+    }).toArray();
 
-    if (modelExists.length) {
+    if (modelExists && modelExists.length) {
       await db.dropCollection(collectionName);
+    } else {
+      console.error(`Model ${modelName} does not exist or could not be found.`);
     }
   } catch (err) {
     throw err;
